Use DynSized type as NonEmptySlice inner type
