+++ tests/compile-fail/boundless-lifetimes/mut.rs
@@ -1,7 +1,7 @@
-use oom::NonEmptyMutSlice;
+use oom::NonEmptySlice;
 fn main() {
     let mut foo = String::from("hello");
-    let mut s = NonEmptyMutSlice::from_mut(&mut foo);
+    let s = NonEmptySlice::from_mut(&mut foo);
     let bad: &'static [_] = s.as_mut_slice();
     drop(foo);
     let _dangling = bad.first();
