+++ tests/compile-fail/boundless-lifetimes/mut2.rs
@@ -1,6 +1,6 @@
 fn main() {
     let mut t = 0;
-    let mut slice = oom::NonEmptyMutSlice::<u8>::from_mut(&mut t);
+    let slice = oom::NonEmptySlice::<u8>::from_mut(&mut t);
     let first = slice.first_mut();
     let second = slice.first_mut();
     assert_eq!(first as *mut _, second as *mut _);
