+++ tests/compile-fail/boundless-lifetimes/vec.stderr
@@ -1,20 +1,11 @@
-error[E0597]: `vec` does not live long enough
- --> $DIR/vec.rs:5:42
+error[E0107]: wrong number of lifetime arguments: expected 0, found 1
+ --> $DIR/vec.rs:5:28
   |
 5 |     let bad: NonEmptySlice<'static, _> = vec.as_nonempty_slice();
-  |              -------------------------   ^^^ borrowed value does not live long enough
-  |              |
-  |              type annotation requires that `vec` is borrowed for `'static`
-...
-8 | }
-  | - `vec` dropped here while still borrowed
+  |                            ^^^^^^^ unexpected lifetime argument
 
-error[E0505]: cannot move out of `vec` because it is borrowed
- --> $DIR/vec.rs:6:10
+error[E0599]: no method named `as_nonempty_slice` found for struct `oom::NonEmptyVec<u8>` in the current scope
+ --> $DIR/vec.rs:5:46
   |
 5 |     let bad: NonEmptySlice<'static, _> = vec.as_nonempty_slice();
-  |              -------------------------   --- borrow of `vec` occurs here
-  |              |
-  |              type annotation requires that `vec` is borrowed for `'static`
-6 |     drop(vec);
-  |          ^^^ move out of `vec` occurs here
+  |                                              ^^^^^^^^^^^^^^^^^ method not found in `oom::NonEmptyVec<u8>`
